+++ inst/examples/showcase/init_showcase.R
@@ -1,6 +1,6 @@
 suppressPackageStartupMessages({
-  #library(ksTFL)
-  devtools::load_all()
+  library(ksTFL)
+  #devtools::load_all()
   library(dplyr)
   library(tidyr)
   library(tibble)